Renamed database util file from sandpit to snowflake
